requirements.txt
# django packages
Django>=4,<5
django-extensions>=3,<4
django-filter==22.1
dj_database_url==0.5.0
djangorestframework>=3,<4
djangorestframework-simplejwt==5.3.1
django-json-widget==2.0.1
drf-standardized-errors==0.13.0
drf-spectacular==0.27.2

# libs
psycopg2-binary>=2,<3
redis>=4,<5
sentry-sdk>=1.5,<2

celery>=5,<6
flower>=1.2,<2
django_celery_beat>=2,<3
django_celery_results>=2,<3
gunicorn==23.0.0
django-ckeditor==6.7.2
